Add an isEmpty() check on chan->GetCurrentInput() in TVRec::SetVideoFiltersForChannel().
It looks like this being null may have caused the problem 
in the partial stack trace Mark Weaver posted today at 12:18pm.

This looks like it is due to some missing initialization code
for handling inputs in DVBChannel, which unfortunately was
not written to be input aware. We don't use video filters on
DVB recordings anyway, so this fix is fine for now.



git-svn-id: http://svn.mythtv.org/trunk@7104 4b91aca1-2a1f-0410-a73a-a78f2df169e8