Partial fix for the "DVB channel flipping hang" reported by Mark Weaver.
It appears this may be a DVB driver problem:
  ioctl(fd_frontend, FE_SET_FRONTEND, &tuning.params)
returned a good value, but the tuner doesn't seem to have changed 
the frequency until much later. This let the signal monitor pick 
up the PAT from the old frequency. But then not get the rest of
the data because the frequency actually did change and the signal
monitor was looking in the wrong place for more data.

The fix is to remove some fallback code that tries to tune to any
PMT if the desired one doesn't exits. The fallback code should no
longer be needed with the signal monitor in place.



git-svn-id: http://svn.mythtv.org/trunk@7125 4b91aca1-2a1f-0410-a73a-a78f2df169e8