Fix for the portion of Stuart Auchterlonie's DVB Scanning problem that he found.
Basically the scanner is reverting to using the PAT to insert
channels, and failing because the NIT in the DB doesn't match
the non-existant NIT of the PAT.

This fixes the cause of the PAT channel insert failing, but 
does not address the underlying problem. Stuart's scanning 
did not find any DVB tables. But I may have already fixed 
this in [7124], so we'll see..



git-svn-id: http://svn.mythtv.org/trunk@7129 4b91aca1-2a1f-0410-a73a-a78f2df169e8