This clamps values set by SignalMonitorValue::SetValue() to values within the SignalMonitorValue range, as set by SetRange().
This fixes the problem that David Peterson and Allan Stirling
were having when they got tables but somehow failed the signal
lock check. This was because the BER was out of the uint16 range,
beause it and UB are actually uint32's that are usually within
the 0-1000 range when actually queried. But they are cumulative
counters so the first value read can be very high.



git-svn-id: http://svn.mythtv.org/trunk@7157 4b91aca1-2a1f-0410-a73a-a78f2df169e8