Fixes #262.
There were essentially two problems. One was that hitting
ESC while the signal OSD was on screen just exited the OSD,
not the player. Then the OSD reappeared once there was a 
new signal message. This is why it appeared to be quiting
but then it didn't quit.

The second problem was that AbortStartRecorderThread() was
called after "QMutexLocker lock(&stateChangeLock)" in 
HandleStateChange(), this meant that if you hit ESC twice
you still didn't exit because the StartRecorderPost() 
needed access to lock the stateChangeLock before it exited.
Since HandleStateChange() wouldn't release the lock until
StartRecorderPost() exited, we had a classic deadlock.

I also fixed two other somewhat related problems:
 1/ HandleStateChange did not reset the changeState to
    false after a Null transition. This created a great
    deal of chatter until you issued next changedStates.
 2/ We did not grab the TV::osdlock before deleting the
    nvp. This presents a problem if we try check GetOSD()
    for null and then use after the nvp is deleted...
    (This is still a problem with anyone using GetOSD()
    outside the TV event loop, without grabbing a
    TV::osdlock. The UDPNotify class comes to mind.)



git-svn-id: http://svn.mythtv.org/trunk@7177 4b91aca1-2a1f-0410-a73a-a78f2df169e8