This should fix the DVB scanning problem reported by John Pullan and others.
The transport was not getting incremented because a temporary
variable had the wrong type. Once the first good transport was
found on a frequency offset other than the last one scanned for
each transport then the transport would loop back to the first
frequency offset on that transport.

This of course only effected scans where multiple frequencies
are scanned for each transport.



git-svn-id: http://svn.mythtv.org/trunk@7185 4b91aca1-2a1f-0410-a73a-a78f2df169e8