Improved backend restart handling.
When a slave connects or disconnects, the scheduler gets notified and
updates its state accordingly.  For example, if a slave disconnects
while it is recording something, the scheduler now knows that
recording was incomplete.  Similarly, if the master restarts while a
slave is recording something, the scheduler now knows that the slave
kept recording.

When the master restarts, it honors the state information saved in the
oldrecorded table.  This should make the master restart case behave
similarly to a slave restart.  Previously, some things would work
differently because the scheduler lost its in-memory state and ignored
any state that was in the database.

Added a new rsAborted recording status for when a backend is stopped
while recording something.  The scheduler will automatically try to
reschedule these recordings.  Note that the rescheduling for aborted
recordings is not as aggressive as the existing reactivate feature.
Reactivating only affects one specific showing.  For an aborted
recording, the scheduler might choose a later (complete) showing
instead of the current one.

Removed the rsStopped and rsDeleted recording statuses.  They added
very little value and complicated the changes described above.  If
anyone really wants them back, I might reconsider, but it will be a
very low priority task.

Handled a tiny race condition where the scheduler tries to start a
recording on a slave right as the slave is disconnecting.  Previously,
the status would have been rsTunerBusy.  Now, the status is rsAborted.


git-svn-id: http://svn.mythtv.org/trunk@7191 4b91aca1-2a1f-0410-a73a-a78f2df169e8