Ref #288, may fix.
This makes the DVBSignalMonitor ignore the DVB Driver's
timeout value. It looks like this may have been causing
some of the problems with multiple cards. My guess is this
made sense when the DVBSignalMonitor was just used for
scanning, but now that we handle timeouts ourselves this
doesn't make any sense.



git-svn-id: http://svn.mythtv.org/trunk@7207 4b91aca1-2a1f-0410-a73a-a78f2df169e8