Fixes #179.
Adds a new functiuon to the NVP ReleaseNextVideoFrame(null)
which gets and releases the next frame that would normally
displayed.

Adds a call to NVP::ReleaseNextVideoFrame(null) in the
AvFormatDecoderPrivate::SeekReset() skipFrames loop.

This should address the XvMC hang in FF x3 mode and
with same problem with exact seeking.



git-svn-id: http://svn.mythtv.org/trunk@7221 4b91aca1-2a1f-0410-a73a-a78f2df169e8