Closes #171.
Adds a new autoexpire policy.

It sorts by priority then age so that lower priority 
shows are deleted first.

This updates the DB schema and I recommend anyone with
a remote backends upgrade them at the same time if they
want this functionality in the future. If you don't any
new recordings on the remote backend will be considered to
have the default priority for the new auto-expire method.



git-svn-id: http://svn.mythtv.org/trunk@7223 4b91aca1-2a1f-0410-a73a-a78f2df169e8