I looked at why UI was so sluggish when the signal monitoring was on screen. Turns out that UpdateOSDSignal was taking much longer to run than I thought it should be. I added some caching for the slower calls which make the UI much more responsive.
I also cleaned up some things, like I made it so that processKey()
debugging can be enabled with by changing a define rather than
changing "#if 0" to "#if 1". This also properly shuts down
a PiP ringbuffer, and sets "wantsToQuit" properly for a few error
exits. It adds some documentation, and times StartRecorder() and
StartPlayer() so I can get some measurements on how long this is
taking for people that complain about speed...



git-svn-id: http://svn.mythtv.org/trunk@7293 4b91aca1-2a1f-0410-a73a-a78f2df169e8