Two refactorings.
First, we replace setup_mpeg_table_monitoring(), setup_atsc_table_monitoring(),
and setup_table_monitoring() with TVRec::SetupTableMonitoring() which
combines these functions with one well documented one.

Second, we get rid of TVRec::SetChannel() which queries the database for
the channel number and input matching curRecording and then changes to
that channel. We replace it with a method in ProgramInfo GetChannel()
that returns the channel number and input for the program, and then
we use that to tune to the channel. This should make implementing #340
a little easier as it means one less dependency to curRecording.



git-svn-id: http://svn.mythtv.org/trunk@7303 4b91aca1-2a1f-0410-a73a-a78f2df169e8