deleteLater() support for recorders.
This is very boilerplate for most of the recorders.

However,

There a problem with the DummyDTVRecorder, its dtor
was leaving it's file descriptor open on shutdown.

The DBox2Recorder dtor however could cause SEGFAULTs
in several places if it's deleted after DBox2Channel()
as it is in ~TVRec. This is because it uses the
deleted channel instance during shutdown. I've
fixed this in the dtor and added initialization
for some uninitialized variables in the ctor.

I cleaned up the FirewireRecorder ctor/dtor, but
nothing major there.



git-svn-id: http://svn.mythtv.org/trunk@7322 4b91aca1-2a1f-0410-a73a-a78f2df169e8