Changes many of the flags in TVRec to use [Set|Clear|Has]Flags().
Then it adds a wait condition which is triggered on
SetFlags(), ClearFlags(), and ChangeState(), this
wait condition lets us increase the sleep time in
the event loop.

This is not a 30 second sleep like in the tuning patch.
We still need to add triggers for the tuning and for
the various busy loops still in TVRec.



git-svn-id: http://svn.mythtv.org/trunk@7330 4b91aca1-2a1f-0410-a73a-a78f2df169e8