Mostly this just gets rid of RingBufferInfo.
But it also adds a SetRingBuffer(NULL) on line 3636,
which gets rid of the RingBuffer if we fail to open
the backing file.

It also gets rid of any dummyRecorder using an old
RingBuffer in SetRingBuffer(). A new one is created
when it is absent and we need one (existing code).



git-svn-id: http://svn.mythtv.org/trunk@7341 4b91aca1-2a1f-0410-a73a-a78f2df169e8