Closes #230 by applying patch.
Basically it fetches the channel_timeout from the DB,
and if it takes longer than that to start getting data
from the backend it pops up an informational dialog.

The change in mythdialog is to add MythMainWindow::GetKey()
which is symmetric to RegisterKey(). It fetches the key
binding from the database for display in informational
dialogs like the above. This lets you tell the user 
exactly what key to press, even if they have remapped
the keys.



git-svn-id: http://svn.mythtv.org/trunk@7344 4b91aca1-2a1f-0410-a73a-a78f2df169e8