This fixes the EIT Scan problem experienced by Mark Edwards.
This skips channels with a blank channum field or with a NULL
mplexid column.

It also clears the pending commands even if we are not in a recording
mode. This should cure any problems when someone abandons LiveTV
before it starts up the recorder, and then the EIT scan kicks in
and tries to start the recorder. I don't think this is actually
a problem because we should take care of this when the recording
is abandoned long before the eit scan starts up, but this simplifies
that code anyway.



git-svn-id: http://svn.mythtv.org/trunk@7350 4b91aca1-2a1f-0410-a73a-a78f2df169e8