Allow entry of ATSC/MPEG channels with normal remote.
This only works if you have smart channel entry enabled or you
are in browse mode. Basically if the channel you entered is not
found it tries adding the spacer, and if that channel is in there
and is unique it will switch to that channel.

So if you have a channel 5-1 (and no channel 51, or 51*),
you can just enter 51, and it will magically fill in the "-".

I also noticed an incompatibility between smart channel entry and
TOGGLECC & arbitrary seek. The problem is that if you have say a 
channel 88 and say want to toggle to the 888 CC table then you will
jump to channel 88 as soon as you've typed those two characters,
if you don't have a channel 88 then it will reset the channelKeys
to 8 as soon as you enter the second 8....

I've partially fixed this by adding an inputKeys buffer in addition
to the existing channelKeys buffer. Then if you say don't have
a channel 90, but want to use that for your arbritary seek,
smart channel entry won't reset you entry to 0 when you type 90.
(but will still reset the channelKeys). If you messed up, you can
still just type a bunch of zeros.



git-svn-id: http://svn.mythtv.org/trunk@7353 4b91aca1-2a1f-0410-a73a-a78f2df169e8