Handle RingBuffer? shutdown better on error conditions.
This seems to have been causing some of the problems
people are reporting. If the StartRecorder() method
failed we didn't enter the kFlagRecorderRunning state,
and hence wouldn't shut down the recorder and ringbuffer.

If for some reason the starting up the recorder failed
in LiveTV we didn't send a QUIT_LIVETV to the frontend,
so StopLiveTV() was never called and so the ringbuffer
for LiveTV remained around. This shouldn't really happen
unless device permissions or compile flags are broken.



git-svn-id: http://svn.mythtv.org/trunk@7367 4b91aca1-2a1f-0410-a73a-a78f2df169e8