Implements a basic "Antenna Adjustment Mode". (F7)
This turns off the recorder in LiveTV to let
you adjust the antenna and see a constant 
signal monitoring OSD.

This could be more useful with a better signal
monitoring OSD, but this is a start.

This may also fix the spurious "You should have
gotten a signal by now" dialogs.



git-svn-id: http://svn.mythtv.org/trunk@7376 4b91aca1-2a1f-0410-a73a-a78f2df169e8