Fix some of the EIT scanning problems people have been seeing.
We add an optional parameter to SetChannel() that lets us specify
that this channel change is in fact for an EIT scan, so that we
don't have to detect this. Also there is no need to wait for a
recorder reset in this case, as there is no recorder...

Also we tell the stream data instance that we don't need video
in the PMT when we are doing an EIT scan.. This lets people in
the UK scan "off-air" channels that are only off-air in the PMT,
but are still transmitting EIT guide data.

Also re-enable passive EIT scan during recordings.



git-svn-id: http://svn.mythtv.org/trunk@7379 4b91aca1-2a1f-0410-a73a-a78f2df169e8