Closes #414.
I didn't use the attached patch.

It turns out the Event class was not being used in a thread-safe way.

This is due to the QString and QStringList classes it contains.
These classes use shared data and either all copies must be locked
or you must make deep copies. I added a deep copy method to Event
and use that to make this safe.

I also changed QList_Events and QListList_Events from QValueList
classes to MythDequeue classes. This lets us convert some of those
loops to simple function calls. And makes sense algorithmically 
since we use them not as vectors but as FIFO queues.



git-svn-id: http://svn.mythtv.org/trunk@7385 4b91aca1-2a1f-0410-a73a-a78f2df169e8