This commit looks a lot bigger then it is.
Close #248 by applying (and fixing) the patch contained therein.  Note
to submitter: Tabs not spaces, when modifiying dbcheck.cpp you need to
bump the schema number and make your changes not change an old schema.

Note: The "video_ts" check was removed it's _always_ "VIDEO_TS"
otherwise it's not a DVD.

This also adds a default play commands for ".iso" to match the one for
"VIDEO_TS".

The remainder of these chnages are const correctness and formatting
changes.



git-svn-id: http://svn.mythtv.org/trunk@7388 4b91aca1-2a1f-0410-a73a-a78f2df169e8