Closes #424 by applying patch.
I did not add the program number change, this would require
also using this program number in the PAT. And there is no
need for recordings to use the same program number.

I also added SetAudioStreamsRequired()/GetAudioStreamsRequired()
to support DVB radio. This defaults to zero, but for DVB radio
you would want it to be at least one.

And I also changed the "delete sd;" in TVRec for DVB with 
sd->deleteLater() for signal/slot safety.



git-svn-id: http://svn.mythtv.org/trunk@7398 4b91aca1-2a1f-0410-a73a-a78f2df169e8