Hack to TV::ToggleRecord which will create a kManualRecord if "R" is pressed when there is no program info. This will create and instantly start a 60 minute manual record from the current minute. However, now that changing the end time of a recording in progress is supported, the endtime can be adjusted at the user's convenience any time after the recording has begun.
The user would be better served by going to the "Manual Schedule" page
where the specific end time can be set as well as other recording
options and the user can enter a title to her liking. Also, this only
applies to TOGGLERECORD in the live TV context. You can't choose to
record something that isn't there in the proglist or progfinder and
choosing to record "Unknown" in the EPG is a bad idea.


git-svn-id: http://svn.mythtv.org/trunk@7401 4b91aca1-2a1f-0410-a73a-a78f2df169e8