Adds RingBuffer switching support to mpegrecorder and dtvrecorder.
There are two new methods:
  RecorderBase::SetNextRecording(const ProgramInfo*, RingBuffer*);
and
  TVRec::RingBufferChanged()

You call the SetNextRecording() to request a ringbuffer and 
program info change, and the recorder will call the 
RingBufferChanged() immediately after the change. The mpegrecorder
and dtvrecorder both wait for the next keyframe and then perform
the switch.

The changes in the other recorders are to add TVRec as a parameter
for the constructors, and in the case of NuppelVideoRecorder to add
a dummy SetNextRecording() method

There is debugging code in TVRec to perform a buffer change,
but it is ifdef'ed out.

The change in programs/mythtranscode/transcode.cpp is simply
to add a NULL TVRec* parameter.

This is in support of the new LiveTV support Issac is working on.



git-svn-id: http://svn.mythtv.org/trunk@7406 4b91aca1-2a1f-0410-a73a-a78f2df169e8