References #430.
Possible fix, this waits for the internal state to change to
kState_None after canceling an overrecord. This isn't entirely
safe because someone may start LiveTV at just the wrong time,
making this block for the length of the LiveTV recording.

But this was an existing problem in the code since overrecord
was first implemented, and no one complained about this actually
happening in practice.



git-svn-id: http://svn.mythtv.org/trunk@7410 4b91aca1-2a1f-0410-a73a-a78f2df169e8